test: 修改@itenl/react-native-scrollable-tabview的demo
