+++ @itenl/react-native-scrollable-tabview/ScrollableTabviewExampleTest/ScrollableTabviewExample19.jsx
@@ -1,0 +1,101 @@
+import React, { useState, useRef } from 'react';
+import { View, Text, ScrollView, TouchableOpacity, Dimensions, StyleSheet, SectionList, StatusBar, Button } from 'react-native'
+import ScrollableTabView, { DefaultTabBar } from '@itenl/react-native-scrollable-tabview';
+import { Tester, TestSuite, TestCase } from '@rnoh/testerino';
+
+class Screen1 extends React.Component {
+    constructor(props) {
+        super(props);
+        this.state = {
+            time: Date.now(),
+        };
+    }
+
+    onRefresh = (toggled) => {
+        this.toggled = toggled;
+        this.toggled && this.toggled();
+        this.toggled && this.toggled();
+    };
+
+    render() {
+        return (
+            <View
+                style={{
+                    flex: 1,
+                    backgroundColor: 'red'
+                }}
+            >
+                <Text>
+                    第一个页面
+                </Text>
+            </View>
+        );
+    }
+}
+
+export default class ScrollableTabviewExample19 extends React.Component {
+    constructor(props) {
+        super(props);
+        this.state = {
+            rootTime: Date.now(),
+            stacks: [],
+            firstIndex: 0,
+            useScroll: false,
+            scroll: "0",
+            scroll2Horizontal: "0",
+            errorText: ''
+        };
+    }
+
+    initStacks() {
+        return [
+            {
+                screen: Screen1,
+                tabLabel: "Screen1",
+            }
+        ];
+    }
+
+    componentDidMount() {
+        this.setState({
+            stacks: this.initStacks(),
+        });
+        setTimeout(() => {
+            const stacks = this.state.stacks;
+            if (stacks && stacks[1]) {
+                stacks[1].tabLabelRender = (tabLabel) => {
+                    return `${tabLabel}`;
+                };
+                this.setState({
+                    stacks,
+                });
+            }
+        }, 5000);
+    }
+
+    componentWillUnmount() {
+        clearTimeout(this.timer);
+    }
+
+    render() {
+        return (
+            <Tester>
+                <TestSuite name="TesterScrollableTabviewExample18">
+                    <TestCase
+                        tags={['C_API']}
+                        itShould="未设置组件高度,设置fillscreen为false,无法填充外层盒子的大小.请点击Screen1查看效果">
+                        <View style={{ width: '100%', height: 500 }}>
+                            <ScrollableTabView
+                                stacks={this.state.stacks}
+                                firstIndex={this.state.firstIndex}
+                                fillScreen={false}
+                            ></ScrollableTabView>
+                        </View>
+                    </TestCase>
+                </TestSuite>
+            </Tester>
+        );
+    }
+}
+
+console.disableYellowBox = true;
