+++ @itenl/react-native-scrollable-tabview/ScrollableTabviewExampleTest/index.tsx
@@ -1,0 +1,146 @@
+
+import { NavigationContainer, Page } from "./navitagiton"
+import ScrollableTabviewExample from './ScrollableTabviewExample';
+import ScrollableTabviewExample2 from './ScrollableTabviewExample2';
+import ScrollableTabviewExample3 from './ScrollableTabviewExample3';
+import ScrollableTabviewExample4 from './ScrollableTabviewExample4';
+import ScrollableTabviewExample5 from './ScrollableTabviewExample5';
+import ScrollableTabviewExample6 from './ScrollableTabviewExample6';
+import ScrollableTabviewExample7 from './ScrollableTabviewExample7';
+import ScrollableTabviewExample8 from './ScrollableTabviewExample8';
+import ScrollableTabviewExample9 from './ScrollableTabviewExample9';
+import ScrollableTabviewExample10 from './ScrollableTabviewExample10';
+import ScrollableTabviewExample11 from './ScrollableTabviewExample11';
+import ScrollableTabviewExample12 from './ScrollableTabviewExample12';
+import ScrollableTabviewExample13 from './ScrollableTabviewExample13';
+import ScrollableTabviewExample14 from './ScrollableTabviewExample14';
+import ScrollableTabviewExample15 from './ScrollableTabviewExample15';
+import ScrollableTabviewExample16 from './ScrollableTabviewExample16';
+import ScrollableTabviewExample17 from './ScrollableTabviewExample17';
+import ScrollableTabviewExample18 from './ScrollableTabviewExample18';
+import ScrollableTabviewExample19 from './ScrollableTabviewExample19';
+import ScrollableTabviewExample20 from './ScrollableTabviewExample20';
+import ScrollableTabviewExample21 from './ScrollableTabviewExample21';
+import ScrollableTabviewExample22 from './ScrollableTabviewExample22';
+import ScrollableTabviewExample23 from './ScrollableTabviewExample23';
+import ScrollableTabviewExample24 from './ScrollableTabviewExample24';
+import ScrollableTabviewExample25 from './ScrollableTabviewExample25';
+import ScrollableTabviewExample26 from './ScrollableTabviewExample26';
+import ScrollableTabviewExample27 from './ScrollableTabviewExample27';
+
+
+export const ScrollableTabviewExampleDemo = () => {
+  return (
+          <NavigationContainer>
+
+            <Page name="ScrollableTabviewExampleDemo">
+              <ScrollableTabviewExample />
+            </Page>
+
+            <Page name="ScrollableTabviewExampleDemo2">
+              <ScrollableTabviewExample2 />
+            </Page>
+            
+            <Page name="ScrollableTabviewExampleDemo3">
+              <ScrollableTabviewExample3 />
+            </Page>
+
+            <Page name="ScrollableTabviewExampleDemo4">
+              <ScrollableTabviewExample4 />
+            </Page>
+
+            <Page name="ScrollableTabviewExampleDemo5">
+              <ScrollableTabviewExample5 />
+            </Page>
+
+            <Page name="ScrollableTabviewExampleDemo6">
+              <ScrollableTabviewExample6 />
+            </Page>
+
+            <Page name="ScrollableTabviewExampleDemo7">
+              <ScrollableTabviewExample7 />
+            </Page>
+
+            <Page name="ScrollableTabviewExampleDemo8">
+              <ScrollableTabviewExample8 />
+            </Page>
+
+            <Page name="ScrollableTabviewExampleDemo9">
+              <ScrollableTabviewExample9 />
+            </Page>
+
+            <Page name="ScrollableTabviewExampleDemo10">
+              <ScrollableTabviewExample10 />
+            </Page>
+
+            <Page name="ScrollableTabviewExampleDemo11">
+              <ScrollableTabviewExample11 />
+            </Page>
+
+            <Page name="ScrollableTabviewExampleDemo12">
+              <ScrollableTabviewExample12 />
+            </Page>
+
+            <Page name="ScrollableTabviewExampleDemo13">
+              <ScrollableTabviewExample13 />
+            </Page>
+
+            <Page name="ScrollableTabviewExampleDemo14">
+              <ScrollableTabviewExample14 />
+            </Page>
+
+            <Page name="ScrollableTabviewExampleDemo15">
+              <ScrollableTabviewExample15 />
+            </Page>
+
+            <Page name="ScrollableTabviewExampleDemo16">
+              <ScrollableTabviewExample16 />
+            </Page>
+
+            <Page name="ScrollableTabviewExampleDemo17">
+              <ScrollableTabviewExample17 />
+            </Page>
+
+            <Page name="ScrollableTabviewExampleDemo18">
+              <ScrollableTabviewExample18 />
+            </Page>
+
+            <Page name="ScrollableTabviewExampleDemo19">
+              <ScrollableTabviewExample19 />
+            </Page>
+
+            <Page name="ScrollableTabviewExampleDemo20">
+              <ScrollableTabviewExample20 />
+            </Page>
+
+            <Page name="ScrollableTabviewExampleDemo21">
+              <ScrollableTabviewExample21 />
+            </Page>
+
+            <Page name="ScrollableTabviewExampleDemo22">
+              <ScrollableTabviewExample22 />
+            </Page>
+
+            <Page name="ScrollableTabviewExampleDemo23">
+              <ScrollableTabviewExample23 />
+            </Page>
+
+            <Page name="ScrollableTabviewExampleDemo24">
+              <ScrollableTabviewExample24 />
+            </Page>
+
+            <Page name="ScrollableTabviewExampleDemo25">
+              <ScrollableTabviewExample25 />
+            </Page>
+
+            <Page name="ScrollableTabviewExampleDemo26">
+              <ScrollableTabviewExample26 />
+            </Page>
+
+            <Page name="ScrollableTabviewExampleDemo27">
+              <ScrollableTabviewExample27 />
+            </Page>
+
+          </NavigationContainer>
+  )
+}
